// apps/event-approval-codeapp/src/generated/models/Paca_approvaldecisionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Paca_approvaldecisionspaca_decisiontypecode = {
  0: 'Approved',
  1: 'Rejected'
} as const;
export type Paca_approvaldecisionspaca_decisiontypecode = keyof typeof Paca_approvaldecisionspaca_decisiontypecode;
export const Paca_approvaldecisionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Paca_approvaldecisionsstatecode = keyof typeof Paca_approvaldecisionsstatecode;
export const Paca_approvaldecisionsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Paca_approvaldecisionsstatuscode = keyof typeof Paca_approvaldecisionsstatuscode;

export interface Paca_approvaldecisionsBase {
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  paca_approvaldecisionid: string;
  paca_approverdisplayname: string;
  paca_approverid: string;
  paca_comment: string;
  paca_decidedat: string;
  paca_decisiontypecode: Paca_approvaldecisionspaca_decisiontypecode;
  "paca_requestid@odata.bind"?: string;
  statecode: Paca_approvaldecisionsstatecode;
  statuscode?: Paca_approvaldecisionsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Paca_approvaldecisions extends Paca_approvaldecisionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  paca_decisiontypecodename?: string;
  paca_requestidname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  paca_requestid?: object;
  _paca_requestid_value?: string;
}